refactor: remove unnecessary parsers/pipe directory
