+++ src/parsers/__tests__/pipe.test.js
